requirements.txt
transformers==4.31.0
datasets==2.13.0
peft==0.4.0
accelerate==0.21.0
bitsandbytes==0.40.2
trl==0.4.7
safetensors>=0.3.1
ipywidgets==7.7.1
huggingface_hub
python-dotenv
